+++ R/cluster_analysis.R
@@ -71,7 +71,9 @@
 #' rez # Show results
 #' predict(rez) # Get clusters
 #' summary(rez) # Extract the centers values (can use 'plot()' on that)
-#' cluster_discrimination(rez) # Perform LDA
+#' if (requireNamespace("MASS", quietly = TRUE)) {
+#'   cluster_discrimination(rez) # Perform LDA
+#' }
 #'
 #' # Hierarchical k-means (more robust k-means)
 #' if (require("factoextra", quietly = TRUE)) {