+++ R/methods_glmmTMB.R
@@ -35,7 +35,12 @@ model_parameters.glmmTMB <- function(model,
   }
 
   # sanity check, warn if unsupported argument is used.
-  dot_args <- .check_dots(dots = list(...), not_allowed = c("vcov", "vcov_args"), class(model)[1], verbose = verbose)
+  dot_args <- .check_dots(
+    dots = list(...),
+    not_allowed = c("vcov", "vcov_args"),
+    class(model)[1],
+    verbose = verbose
+  )
 
   # p-values, CI and se might be based on different df-methods
   ci_method <- .check_df_method(ci_method)
@@ -124,7 +129,18 @@ model_parameters.glmmTMB <- function(model,
         params[nrow(params), "Parameter"] <- dispersion_param$Parameter[1]
         params[nrow(params), "Coefficient"] <- stats::sigma(model)
         params[nrow(params), "Component"] <- dispersion_param$Component[1]
-        params[nrow(params), c("CI_low", "CI_high")] <- stats::confint(model, parm = "sigma", method = "wald", level = ci)[1:2]
+        params[nrow(params), c("CI_low", "CI_high")] <- tryCatch(
+          suppressWarnings(stats::confint(model, parm = "sigma", method = "wald", level = ci)[1:2]),
+          error = function(e) {
+            if (verbose) {
+              message(insight::format_message(
+                "Cannot compute standard errors and confidence intervals for sigma parameter.",
+                "Your model may suffer from singularity (see '?lme4::isSingular' and '?performance::check_singularity')."
+              ))
+            }
+            c(NA, NA)
+          }
+        )
         dispersion_param <- TRUE
       }
     }
@@ -241,6 +257,15 @@ ci.glmmTMB <- function(x,
     return(NULL)
   }
 
+  # sanity check, warn if unsupported argument is used.
+  dot_args <- .check_dots(
+    dots = list(...),
+    not_allowed = c("vcov", "vcov_args"),
+    class(x)[1],
+    function_name = "ci",
+    verbose = verbose
+  )
+
   # profiled CIs
   if (method == "profile") {
     pp <- stats::profile(x)
@@ -265,12 +290,20 @@ ci.glmmTMB <- function(x,
 #' @rdname standard_error
 #' @export
 standard_error.glmmTMB <- function(model,
-                                   effects = c("fixed", "random"),
-                                   component = c("all", "conditional", "zi", "zero_inflated", "dispersion"),
+                                   effects = "fixed",
+                                   component = "all",
                                    verbose = TRUE,
                                    ...) {
-  component <- match.arg(component)
-  effects <- match.arg(effects)
+  component <- match.arg(component, choices = c("all", "conditional", "zi", "zero_inflated", "dispersion"))
+  effects <- match.arg(effects, choices = c("fixed", "random"))
+
+  dot_args <- .check_dots(
+    dots = list(...),
+    not_allowed = c("vcov", "vcov_args"),
+    class(model)[1],
+    function_name = "standard_error",
+    verbose = verbose
+  )
 
   if (effects == "random") {
     if (requireNamespace("TMB", quietly = TRUE) && requireNamespace("glmmTMB", quietly = TRUE)) {