+++ R/methods_mjoint.R
@@ -35,7 +35,7 @@ model_parameters.mjoint <- function(model,
   }
 
   if (effects %in% c("random", "all")) {
-    params_variance <- .extract_random_variances(model, ci = ci, effects = effects, ci_method = NULL)
+    params_variance <- .extract_random_variances(model, ci = ci, effects = effects, ci_method = NULL, verbose = verbose)
     params_variance$Component <- "conditional"
   }