+++ NEWS.md
@@ -1,4 +1,16 @@
-# parameters 0.17.1
+# parameters 0.18.1
+
+## Changes to functions
+
+* `model_parameters()` now also includes standard errors and confidence
+  intervals for slope-slope-correlations of random effects variances.
+
+## Bug fixes
+
+* Fixed issue with *glmmTMB* models when calculating confidence 
+  intervals for random effects failed due to singular fits.
+
+# parameters 0.18.0
 
 ## Breaking Changes